Marshal: compiler warning
git-svn-id: file:///home/michael/svn/yuni/trunk@2333 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/marshal/object.cpp
@@ -219,9 +219,9 @@ namespace Marshal
 		switch (pType)
 		{
 			case otArray:
-				return pValue.array->size();
+				return (uint )pValue.array->size();
 			case otDictionary:
-				return pValue.dictionary->size();
+				return (uint) pValue.dictionary->size();
 			case otNil:
 				return 0;
 			default:
